+++ crypto-primitives/benches/merkle_tree.rs
@@ -4,9 +4,9 @@ extern crate criterion;
 static NUM_LEAVES: i32 = 1 << 20;
 
 mod bytes_mt_benches {
-    use ark_crypto_primitives::crh::*;
-    use ark_crypto_primitives::merkle_tree::*;
-    use ark_crypto_primitives::to_uncompressed_bytes;
+    use crate::crh::*;
+    use crate::merkle_tree::*;
+    use crate::to_uncompressed_bytes;
     use ark_ff::BigInteger256;
     use ark_serialize::CanonicalSerialize;
     use ark_std::{test_rng, UniformRand};