+++ crypto-primitives/src/crh/bowe_hopwood/mod.rs
@@ -234,8 +234,8 @@ impl<P: TECurveConfig, W: pedersen::Window> TwoToOneCRHScheme for TwoToOneCRH<P,
     ) -> Result<Self::Output, Error> {
         Self::evaluate(
             parameters,
-            ark_crypto_primitives::to_uncompressed_bytes!(left_input)?,
-            ark_crypto_primitives::to_uncompressed_bytes!(right_input)?,
+            crate::to_uncompressed_bytes!(left_input)?,
+            crate::to_uncompressed_bytes!(right_input)?,
         )
     }
 }